src/world_model/train_world_model.py
import argparse
import yaml
import os
import torch
from src.world_model.models import CtrlWorldModel
from src.envs.droid import get_droid_dataset

def main():
    parser = argparse.ArgumentParser()
    parser.add_argument("--config", type=str, required=True, help="Path to config yaml")
    parser.add_argument("--exp_name", type=str, default="ctrl_world_droid", help="Experiment name")
    args = parser.parse_args()

    with open(args.config, "r") as f:
        config = yaml.safe_load(f)

    wm_config = config.get("world_model", {})
    checkpoint_dir = wm_config.get("checkpoint_dir", "yjguo/Ctrl-World")
    steps = wm_config.get("steps", 50000)
    batch_size = wm_config.get("batch_size", 8)

    # DROID dummy dataloader
    dataset = get_droid_dataset()
    
    # Load model
    model = CtrlWorldModel(checkpoint_dir=checkpoint_dir)
    optimizer = torch.optim.Adam(model.parameters(), lr=1e-4)

    print(f"Starting world model fine-tuning for {steps} steps on DROID data...")
    step = 0
    
    # Dummy Training Loop
    while step < steps:
        for batch in dataset:
            optimizer.zero_grad()
            
            # Predict future frame based on history and actions
            # This is a stub for the diffusion loss computation
            loss = torch.tensor(0.5, requires_grad=True)
            loss.backward()
            optimizer.step()
            
            if step % 1000 == 0:
                print(f"Step {step}/{steps} - Loss: {loss.item():.4f}")
            
            step += 1
            if step >= steps:
                break
                
    print("Training finished.")

if __name__ == "__main__":
    main()
